requirements.txt
requests==2.10.0
lxml
python-qbittorrent